+++ specs/src/main/java/org/seedstack/seed/spi/SeedInitializer.java
@@ -10,19 +10,35 @@ package org.seedstack.seed.spi;
 import org.seedstack.coffig.Coffig;
 
 /**
- * This interface defines two methods that are called at Seed JVM initialization and close.
- * Implementations must be declared as a {@link java.util.ServiceLoader} service in META-INF/services to be detected.
+ * <p>This interface defines methods that are called at various stages of the SeedStack JVM initialization and shutdown
+ * process. Implementations must be declared as a {@link java.util.ServiceLoader} service in META-INF/services to be
+ * detected.</p>
+ *
+ * <p>Classes implementing this interface can be annotated with {@link javax.annotation.Priority} to specify an
+ * absolute order among them.</p>
+ *
+ * <p>A single instance of each implementation is created and using throughout the whole lifecycle.</p>
  */
 public interface SeedInitializer {
     /**
-     * Called at Seed JVM-wide initialization.
+     * Called before SeedStack initialization.
+     */
+    void beforeInitialization();
+
+    /**
+     * Called during Seed initialization, just after base configuration has been made available.
      *
      * @param configuration the base configuration.
      */
     void onInitialization(Coffig configuration);
 
     /**
-     * Called at Seed JVM-wide close.
+     * Called after Seed initialization has been completed.
+     */
+    void afterInitialization();
+
+    /**
+     * Called at explicit Seed global state cleanup.
      */
     void onClose();
 }