+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Version 3.3.1 (2017-09-15)
 
 * [new] Configuration dump (config tool) now dumps inner properties for maps, collections, arrays and complex objects.
-* [new] Add `isRemembered()` on `SecuritySupport`.
+* [new] Add `beforeInitialization()` and `afterInitialization()` methods on `SeedInitializer` interface.
+* [new] Add `isRemembered()` on `SecuritySupport` interface.
  
 # Version 3.3.0 (2017-07-31)